+++ rice-layout/src/rect.rs
@@ -3,8 +3,8 @@
 /// A rectangle with position and size
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub struct Rect {
-    pub x: usize,
-    pub y: usize,
-    pub height: usize,
-    pub width: usize,
+    pub x: i32,
+    pub y: i32,
+    pub height: i32,
+    pub width: i32,
 }
